#270 rename all JS imports to use filenames
