+++ lib/package.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.34.1",
+  "version": "0.34.5",
   "author": "Joep Meindertsma",
   "dependencies": {
     "@noble/ed25519": "1.6.0",
