+++ lib/src/commit.ts
@@ -6,8 +6,6 @@ import { decode as decodeB64, encode as encodeB64 } from 'base64-arraybuffer';
 import { sha512 } from '@noble/hashes/sha512';
 utils.sha512 = msg => Promise.resolve(sha512(msg));
 
-import { properties, urls } from './urls';
-import { Store } from './store';
 import {
   isArray,
   JSONArray,
@@ -15,7 +13,10 @@ import {
   parseJsonAdResourceValue,
   removeQueryParamsFromURL,
   Resource,
-} from './index';
+  properties,
+  urls,
+  Store,
+} from './index.js';
 
 /** A {@link Commit} without its signature, signer and timestamp */
 export interface CommitBuilderI {
