+++ react/package.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.34.1",
+  "version": "0.34.8",
   "author": "Joep Meindertsma",
   "description": "Atomic Data React library",
   "dependencies": {
