+++ react/src/index.ts
@@ -22,12 +22,12 @@
  * @module
  */
 
-export * from './hooks';
-export * from './useServerURL';
-export * from './useCurrentAgent';
-export * from './useDebounce';
-export * from './useImporter';
-export * from './useLocalStorage';
-export * from './useMarkdown';
-export * from './useServerSearch';
+export * from './hooks.js';
+export * from './useServerURL.js';
+export * from './useCurrentAgent.js';
+export * from './useDebounce.js';
+export * from './useImporter.js';
+export * from './useLocalStorage.js';
+export * from './useMarkdown.js';
+export * from './useServerSearch.js';
 export * from '@tomic/lib';
