+++ react/src/useMarkdown.ts
@@ -9,7 +9,7 @@ import {
   valToDate,
 } from '@tomic/lib';
 import { useEffect, useState } from 'react';
-import { useStore, useString, useTitle } from './index';
+import { useStore, useString, useTitle } from './index.js';
 
 /** Properties not relevant to show in PropValLines */
 const hiddenProps = [
