+++ react/src/useServerSearch.tsx
@@ -1,6 +1,6 @@
 import { urls } from '@tomic/lib';
 import { useEffect, useMemo, useState } from 'react';
-import { useArray, useDebounce, useResource, useStore } from './index';
+import { useArray, useDebounce, useResource, useStore } from './index.js';
 
 interface SearchResults {
   /** Subject URLs for resources that match the query */
