+++ tsconfig.build.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "ES5",
-    "module": "commonjs",
+    "module": "ES2022",
     "lib": [
       "ES6",
       "ES7",
